--- sample/.gitignore
@@ -1,22 +1,0 @@
-*.swp
-package-lock.json
-.pytest_cache
-*.egg-info
-
-# Byte-compiled / optimized / DLL files
-__pycache__/
-*.py[cod]
-*$py.class
-
-# Environments
-.env
-.venv
-env/
-venv/
-ENV/
-env.bak/
-venv.bak/
-
-# CDK Context & Staging files
-.cdk.staging/
-cdk.out/
