+++ sample/sample.egg-info/requires.txt
@@ -1,0 +1,6 @@
+aws-cdk.core==1.122.0
+aws-cdk.aws_iam==1.122.0
+aws-cdk.aws_sqs==1.122.0
+aws-cdk.aws_sns==1.122.0
+aws-cdk.aws_sns_subscriptions==1.122.0
+aws-cdk.aws_s3==1.122.0
